Update client roles on user object when creating a new account using social auth
