+++ src/routes/user.js
@@ -4,10 +4,10 @@ import bcrypt from 'bcryptjs';
 import {
     createJWTCookie,
     sendVerificationEmail,
-    assignRoleToUsers,
+    addRoles,
 } from '../utils/utils';
 import { refreshTokenName } from '../config/keys';
-import { User, Role } from '../models/user';
+import { User } from '../models/user';
 
 const router = express.Router();
 
@@ -141,15 +141,6 @@ router.post('/register', async (req, res) => {
         // Save the updated the user in database
         await user.save();
 
-        // eslint-disable-next-line no-shadow
-        const addRoles = async (user) => {
-            const roles = await Role.find({});
-            for (let index = 0; index < roles.length; index += 1) {
-                const role = roles[index];
-                // eslint-disable-next-line no-await-in-loop
-                await assignRoleToUsers(role, false, [user]);
-            }
-        };
         addRoles(user);
         sendVerificationEmail(user);
         res.render('register', {
